+++ src/database/repositories/restaurantOrderRepository.ts
@@ -1,0 +1,14 @@
+const getAllOrdersFromRestaurant = () => {};
+
+const getOneOrderFromRestaurant = () => {};
+
+const updateOrderStatusFromRestaurant = () => {};
+
+const cancelOrderFromRestaurant = () => {};
+
+export default {
+    getAllOrdersFromRestaurant,
+    getOneOrderFromRestaurant,
+    updateOrderStatusFromRestaurant,
+    cancelOrderFromRestaurant,
+  };
